requirements.txt
apscheduler==3.11.2
bcrypt==4.0.1
fastapi==0.135.3
fastapi-cache2==0.2.2
httpx==0.28.1
networkx==3.6.1
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
redis==4.6.0
requests==2.33.1
sendgrid==6.12.5
stripe==15.0.1
uvicorn[standard]==0.44.0
